requirements.txt
fastapi==0.110.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.4
uvicorn==0.29.0
python-jose[cryptography]==3.3.0  # Для JWT
passlib[bcrypt]==1.7.4           # Для хеширования паролей
python-multipart==0.0.9          # Для обработки форм
httpx==0.27.0                    # Для HTTP-запросов (OAuth)
authlib==1.3.0                   # Для OAuth (Google)